6_demo_local_RAG_project/requirements.txt
python-dotenv
langchain
langchain-community
faiss-cpu
ollama
langchain-ollama
marimo
unstructured
markdown
